requirements.txt
streamlit==0.64.0
Keras==2.4.3
tensorflow-cpu==2.3.0
numpy==1.18.4
opencv_python==4.3.0.36
Pillow==7.2.0
